fix: wxWidgets format string assertion - use %d instead of %zu for size_t
Fixed assertion failure in wxArgNormalizer when using wxString::Format
with %zu format specifier on MinGW-w64.

Changes:
- ProxyListPanel.cpp:92 - use %d with static_cast<int>() for row numbers
- SubscriptionPanel.cpp:98 - use %d with static_cast<int>() for proxy counts

This ensures compatibility with wxWidgets strict format checking.

+++ src/ui/SubscriptionPanel.cpp
@@ -95,7 +95,7 @@ void SubscriptionPanel::loadSubscriptions() {
 
         // Count proxies for this subscription
         auto proxies = controller_->loadProxies(sub.id);
-        row.push_back(wxVariant(wxString::Format("%zu", proxies.size())));
+        row.push_back(wxVariant(wxString::Format("%d", static_cast<int>(proxies.size()))));
         row.push_back(wxVariant(sub.updatetime));
         store_->AppendItem(row);
     }
